{{cookiecutter.project_dir}}/requirements.txt
pytz
django-service-urls
django{% if cookiecutter.project_type == "django-cms" %} < 2.0.0{% endif %}
django-fluo >= 0.4.0a5{% if cookiecutter.use_gunicorn == "y" %}
gunicorn{% endif %}{% if cookiecutter.use_cookielaw == "y" %}
django-fluo-legal{% endif %}{% if cookiecutter.use_whitenoise == "y" %}
whitenoise{% endif %}{% if cookiecutter.use_sekizai == "y" or cookiecutter.project_type == "django-cms" %}
django-sekizai{% endif %}{% if cookiecutter.use_widget_tweaks == "y" or cookiecutter.project_type == "django-cms" %}
django-widget-tweaks{% endif %}{% if cookiecutter.use_sorl_thumbnail == "y" %}
sorl-thumbnail{% endif %}{% if cookiecutter.use_djangorestframework == "y" %}

# django-rest-framework
djangorestframework{% endif %}{% if cookiecutter.project_type == "django-cms" %}

# django-cms
django-cms
django-filer
djangocms-column
djangocms-file
djangocms-googlemap
djangocms-link
djangocms-picture
djangocms-style
djangocms-teaser
djangocms-text-ckeditor
djangocms-video
django-polymorphic
django-select2{% endif %}
